添加了configserver 添加了Feign支持 添加了Zuul支持

+++ zhangli-service/src/main/java/chenyunlong/zhangli/exception/MyException.java
@@ -11,9 +11,9 @@ import lombok.NoArgsConstructor;
 public class MyException extends Exception {
 
     private String msg;
-    private Long code;
+    private int code;
 
-    public MyException(String msg, Long code) {
+    public MyException(String msg, int code) {
         super(msg);
         this.msg = msg;
         this.code = code;
